app/requirements.txt
Django==4.2.6
mysqlclient==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
